requirements.txt
Pillow==9.0.1
PixivPy==3.7.0
pyTelegramBotAPI==4.4.0
pytesseract==0.3.9
requests==2.27.1
selenium==4.1.3
telegram==0.0.1
